fix: resolve all cppcheck warnings in chmod, less, tr, wsim
- chmod: fscanf %o requires unsigned int* — changed m to unsigned int
- less: reorder while condition to check bounds before array index
- tr: initialize comp[256] to suppress false-positive uninitvar warning
- wsim: split tok[i]=norm[i++] — undefined evaluation order (UB fix)

+++ src/coreutils/chmod.c
@@ -22,10 +22,10 @@ static int read_meta_mode(const char *path, int *mode) {
     meta_path(path, mp, sizeof(mp));
     FILE *f = fopen(mp, "r");
     if (!f) return 0;
-    int m = 0;
+    unsigned int m = 0;
     int ok = fscanf(f, "%o", &m) == 1;
     fclose(f);
-    if (ok) { *mode = m; return 1; }
+    if (ok) { *mode = (int)m; return 1; }
     return 0;
 }
